crosslang-taskflow-kafka/crosslang-taskflow/auth_service/requirements.txt
flask==3.0.3
flask-bcrypt==1.0.1
flask-cors==4.0.0
pymongo==4.7.2
python-dotenv==1.0.1
PyJWT==2.8.0
